requirements.txt
streamlit
pandas
openpyxl
email
google-adk
